[mdscraper] update for pipenv
