+++ python/src/mdscraper/md_scrape.py
@@ -1,20 +1,7 @@
-#!/usr/bin/env python3
-
-import networkx as nx
 from iok.meta import KnowledgeGraph
-from iok.types import ResourceType, NodeType
+from iok.types import ResourceType
 import re
-import requests
-import matplotlib.pyplot as plt
 from hashlib import sha256
-import json
-import sys
-import uuid
-from networkx.readwrite import json_graph
-
-# configs
-JSON_FILE = "graph.json"
-GRAPH_FILE = "graph.png"
 
 # regex for ordering lines based on hierarchy
 HIERARCHY = [r"^[*-]", r"^[A-Za-z]", r"^###", r"^##", r"^#"]
@@ -151,37 +138,3 @@ def match_line(match: str, line: str, scopes: list, graph: KnowledgeGraph) -> li
             return match_line(match, line, scopes_cpy, graph)
         except IndexError:
             return scopes_cpy
-
-
-# returns the graph for debugging
-def main(link: str) -> KnowledgeGraph:
-
-    f = requests.get(link)
-    text = f.text
-    textlines = text.splitlines()
-
-    scopes = []
-    graph = KnowledgeGraph(0)
-    for line in textlines:
-        match = match_hierarchy(line)
-        scopes = match_line(match, line, scopes, graph)
-
-    return graph
-
-
-if __name__ == "__main__":
-    if len(sys.argv) < 2:
-        print("Missing link argument")
-        print("Expected: md_scrape <raw md link>")
-        print(
-            "Example: md_scrape https://raw.githubusercontent.com/rustielin/iok/master/README.md"
-        )
-        exit(1)
-
-    link = sys.argv[1]
-    g = main(link)
-
-    # draw to file for debugging
-    g.write_graph(GRAPH_FILE)
-
-    g.write_to_json(JSON_FILE)
