--- python/src/mdscraper/requirements.txt
@@ -1,14 +1,0 @@
-certifi==2020.4.5.1
-chardet==3.0.4
-cycler==0.10.0
-decorator==4.4.2
-idna==2.9
-kiwisolver==1.2.0
-matplotlib==3.2.1
-networkx==2.4
-numpy==1.18.3
-pyparsing==2.4.7
-python-dateutil==2.8.1
-requests==2.23.0
-six==1.14.0
-urllib3==1.25.9
